+++ server/common/Constants.ts
@@ -1,0 +1,7 @@
+export const httpStatus = {
+    okay: 200,
+    badRequest: 400,
+    unAuthorize: 401,
+    notFound: 404,
+    internalServerError: 500
+};
