+++ server/validators/interfaces/RequestValidator.ts
@@ -1,0 +1,7 @@
+import { Response, Request } from "express";
+
+interface RequestValidator {
+    validate(req: Request, res: Response, next): void;
+}
+
+export default RequestValidator;
